# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: C7NTAX API
  description: |
    Professional Services Automation API — programmatic access to the C7NTAX platform.
    All endpoints require Bearer token authentication (JWT) obtained via `/api/auth/login`.
    External apps and user-created plugins use this same API.

    ## Authentication
    1. POST `/api/auth/login` with email + password
    2. If MFA is enabled, receive `mfaToken` and POST `/api/auth/verify-mfa` with TOTP code
    3. Use returned `token` as `Authorization: Bearer <token>` header

    ## Rate Limiting
    100 requests per 15-minute window per token. 429 responses include `Retry-After`.

    ## Permissions
    Each endpoint lists required permission(s). Client users are scoped to their company's data.
  version: "1.0.0"
  contact:
    name: Cyber 7 Group
    email: support@cyber7group.com
servers:
  - url: http://localhost:4000/api
    description: Local development
  - url: https://api.C7NTAX.com/api
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all users
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: role, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/User" } }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
    post:
      tags: [Users]
      summary: Create a new user (admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, role]
              properties:
                email: { type: string }
                password: { type: string }
                firstName: { type: string }
                lastName: { type: string }
                role: { $ref: "#/components/schemas/SystemRole" }
                companyId: { type: string }
      responses:
        "201": { description: "User created", content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
  /users/me:
    get:
      tags: [Users]
      summary: Get current authenticated user profile
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: "Current user profile", content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
    patch:
      tags: [Users]
      summary: Update user
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { application/json: { schema: { type: object } } } }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
    delete:
      tags: [Users]
      summary: Deactivate user (soft delete)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: "User deactivated" }

  # ─── Tickets ───
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SystemRole:
      type: string
      enum: [admin, manager, technician, dispatcher, billing_manager, client_admin, client_user, read_only]
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        role: { $ref: "#/components/schemas/SystemRole" }
        status: { type: string }
        mfaEnabled: { type: boolean }
        companyId: { type: string }
        lastLoginAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
...